+++ src-ui/app/views/dashboard-view/views/shutdown-automations-view/tabs/shutdown-automations-triggers-tab/shutdown-automations-triggers-tab.component.ts
@@ -1,0 +1,200 @@
+import { Component, DestroyRef, OnInit } from '@angular/core';
+import {
+  AUTOMATION_CONFIGS_DEFAULT,
+  ShutdownAutomationsConfig,
+} from '../../../../../../models/automations';
+import { cloneDeep } from 'lodash';
+import { AutomationConfigService } from '../../../../../../services/automation-config.service';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
+import { fade, vshrink } from '../../../../../../utils/animations';
+
+@Component({
+  selector: 'app-shutdown-automations-triggers-tab',
+  templateUrl: './shutdown-automations-triggers-tab.component.html',
+  styleUrls: ['./shutdown-automations-triggers-tab.component.scss'],
+  animations: [fade(), vshrink()],
+})
+export class ShutdownAutomationsTriggersTabComponent implements OnInit {
+  protected config: ShutdownAutomationsConfig = cloneDeep(
+    AUTOMATION_CONFIGS_DEFAULT.SHUTDOWN_AUTOMATIONS
+  );
+  protected onSleepActivationWindowStart = '00:00';
+  protected onSleepActivationWindowEnd = '00:00';
+  protected whenAloneActivationWindowStart = '00:00';
+  protected whenAloneActivationWindowEnd = '00:00';
+  protected onSleepDurationString = '00:00:00';
+  protected whenAloneDurationString = '00:00:00';
+
+  constructor(private automationConfigs: AutomationConfigService, private destroyRef: DestroyRef) {}
+
+  ngOnInit() {
+    this.automationConfigs.configs
+      .pipe(takeUntilDestroyed(this.destroyRef))
+      .subscribe((configs) => {
+        this.config = configs.SHUTDOWN_AUTOMATIONS;
+        this.onSleepDurationString = this.durationToString(this.config.triggerOnSleepDuration);
+        this.onSleepActivationWindowStart = this.config.triggerOnSleepActivationWindowStart
+          .map((v) => v.toString().padStart(2, '0'))
+          .join(':');
+        this.onSleepActivationWindowEnd = this.config.triggerOnSleepActivationWindowEnd
+          .map((v) => v.toString().padStart(2, '0'))
+          .join(':');
+        this.whenAloneDurationString = this.durationToString(this.config.triggerWhenAloneDuration);
+        this.whenAloneActivationWindowStart = this.config.triggerWhenAloneActivationWindowStart
+          .map((v) => v.toString().padStart(2, '0'))
+          .join(':');
+        this.whenAloneActivationWindowEnd = this.config.triggerWhenAloneActivationWindowEnd
+          .map((v) => v.toString().padStart(2, '0'))
+          .join(':');
+      });
+  }
+
+  async onChangeOnSleepActivationWindowStart(value: string) {
+    const parsedValue = value
+      .split(':')
+      .map((v) => parseInt(v))
+      .slice(0, 2) as [number, number];
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerOnSleepActivationWindowStart: parsedValue,
+      }
+    );
+  }
+
+  async onChangeOnSleepActivationWindowEnd(value: string) {
+    const parsedValue = value
+      .split(':')
+      .map((v) => parseInt(v))
+      .slice(0, 2) as [number, number];
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerOnSleepActivationWindowEnd: parsedValue,
+      }
+    );
+  }
+
+  async onSleepDurationChange(value: string) {
+    let [hours, minutes, seconds] = value.split(':').map((v) => parseInt(v));
+    if (isNaN(hours)) hours = 0;
+    if (isNaN(minutes)) minutes = 0;
+    if (isNaN(seconds)) seconds = 0;
+    const duration = (hours * 3600 + minutes * 60 + seconds) * 1000;
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerOnSleepDuration: duration,
+      }
+    );
+  }
+
+  async whenAloneDurationChange(value: string) {
+    let [hours, minutes, seconds] = value.split(':').map((v) => parseInt(v));
+    if (isNaN(hours)) hours = 0;
+    if (isNaN(minutes)) minutes = 0;
+    if (isNaN(seconds)) seconds = 0;
+    const duration = (hours * 3600 + minutes * 60 + seconds) * 1000;
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerWhenAloneDuration: duration,
+      }
+    );
+  }
+
+  async onChangeWhenAloneActivationWindowStart(value: string) {
+    const parsedValue = value
+      .split(':')
+      .map((v) => parseInt(v))
+      .slice(0, 2) as [number, number];
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerWhenAloneActivationWindowStart: parsedValue,
+      }
+    );
+  }
+
+  async onChangeWhenAloneActivationWindowEnd(value: string) {
+    const parsedValue = value
+      .split(':')
+      .map((v) => parseInt(v))
+      .slice(0, 2) as [number, number];
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerWhenAloneActivationWindowEnd: parsedValue,
+      }
+    );
+  }
+
+  async toggleTriggerOnSleep() {
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerOnSleep: !this.config.triggerOnSleep,
+      }
+    );
+  }
+
+  async toggleOnSleepActivationWindow() {
+    // Toggle the activation window
+    const config: Partial<ShutdownAutomationsConfig> = {
+      triggerOnSleepActivationWindow: !this.config.triggerOnSleepActivationWindow,
+    };
+    // Reset the window back to default when turning off the activation window
+    if (!config.triggerOnSleepActivationWindow) {
+      config.triggerOnSleepActivationWindowStart = cloneDeep(
+        AUTOMATION_CONFIGS_DEFAULT.SHUTDOWN_AUTOMATIONS.triggerOnSleepActivationWindowStart
+      );
+      config.triggerOnSleepActivationWindowEnd = cloneDeep(
+        AUTOMATION_CONFIGS_DEFAULT.SHUTDOWN_AUTOMATIONS.triggerOnSleepActivationWindowEnd
+      );
+    }
+    // Apply & Save
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      config
+    );
+  }
+
+  async toggleTriggerWhenAlone() {
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      {
+        triggerWhenAlone: !this.config.triggerWhenAlone,
+      }
+    );
+  }
+
+  async toggleWhenAloneActivationWindow() {
+    // Toggle the activation window
+    const config: Partial<ShutdownAutomationsConfig> = {
+      triggerWhenAloneActivationWindow: !this.config.triggerWhenAloneActivationWindow,
+    };
+    // Reset the window back to default when turning off the activation window
+    if (!config.triggerWhenAloneActivationWindow) {
+      config.triggerWhenAloneActivationWindowStart = cloneDeep(
+        AUTOMATION_CONFIGS_DEFAULT.SHUTDOWN_AUTOMATIONS.triggerWhenAloneActivationWindowStart
+      );
+      config.triggerWhenAloneActivationWindowEnd = cloneDeep(
+        AUTOMATION_CONFIGS_DEFAULT.SHUTDOWN_AUTOMATIONS.triggerWhenAloneActivationWindowEnd
+      );
+    }
+    // Apply & Save
+    await this.automationConfigs.updateAutomationConfig<ShutdownAutomationsConfig>(
+      'SHUTDOWN_AUTOMATIONS',
+      config
+    );
+  }
+
+  private durationToString(sleepDuration: number): string {
+    const hours = Math.floor(sleepDuration / 3600000);
+    const minutes = Math.floor((sleepDuration % 3600000) / 60000);
+    const seconds = Math.floor((sleepDuration % 60000) / 1000);
+    return `${hours.toString().padStart(2, '0')}:${minutes.toString().padStart(2, '0')}:${seconds
+      .toString()
+      .padStart(2, '0')}`;
+  }
+}