+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Additional configuration options for the sleep detector (controller presence, sleeping pose)
 - Options for forcing the power state of base stations (Right click the power button).
 - Support for base stations that don't report their status. (Newer lighthouses sold through Valve, likely manufactured by HTC)
+- Automation trigger for the shutdown sequence to run when you've been left alone in your VRChat world instance for a certain amount of time.
 - Custom window titlebar
 
 ### Changed